add version history to edit secrets script

+++ Edit-Secrets.ps1
@@ -1,7 +1,11 @@
 param (
     [string]$File = 'Secrets.json',
-    [string]$RepositoryName = ((git remote get-url origin).Split("/")[-1].Replace(".git","")),
-    [string]$SetFile = 'Set-Secrets.ps1'
+    [string]$RepositoryName = (git remote get-url origin).Split("/")[-1].Replace(".git",""),
+    [string]$SetFile = 'Set-Secrets.ps1',
+    [string]$KeyVaultName,
+    [string]$SecretName,
+    [switch]$VersionHistory = $false,
+    [int]$VersionHistoryLength = 10
 )
 
 # Check to see if Azure PowerShell Module is installed
@@ -10,16 +14,23 @@ if (!(Get-Module -ListAvailable Az.KeyVault)) {
     Install-Module -Name Az.KeyVault -Confirm:$false
 }
 
-Clear-Content -Path "${File}.tmp" -ErrorAction SilentlyContinue
+if (!$VersionHistory) {
+    Clear-Content -Path "${File}.tmp" -ErrorAction SilentlyContinue
+}
 
-Write-Host "Searching for key vaults with tag: 'repository-name=$RepositoryName'" -ForegroundColor DarkGray
-$KeyVaultNames = (Get-AzKeyVault -Tag @{"repository-name" = "$RepositoryName" }).VaultName
+if (!$PSBoundParameters.ContainsKey('KeyVaultName')) {
+    Write-Host "Searching for key vaults with tag: 'repository-name=$RepositoryName'" -ForegroundColor DarkGray
+    $KeyVaultNames = (Get-AzKeyVault -Tag @{"repository-name" = "$RepositoryName" }).VaultName
 
-if ($KeyVaultNames) {
-    Write-Host "Key vaults found: $KeyVaultNames" -ForegroundColor DarkGray
+    if ($KeyVaultNames) {
+        Write-Host "Key vaults found: $KeyVaultNames" -ForegroundColor DarkGray
+    }
+    else {
+        Write-Error "No key vaults found. Please make sure the key vault is tagged correctly"  -ForegroundColor Red
+    }
 }
 else {
-    Write-Error "No key vaults found. Please make sure the key vault is tagged correctly"  -ForegroundColor Red
+    $KeyVaultNames = $KeyVaultName
 }
 
 $KeyVaults = New-Object PSCustomObject
@@ -27,20 +38,55 @@ $KeyVaultNames | ForEach-Object {
     $KeyVaultName = $_
     Write-Host "Generating secrets for $KeyVaultName..." -ForegroundColor DarkGray
     $KeyVault = @()
-    $Secrets = (Get-AzKeyVaultSecret -VaultName $_).Name
+    if ($PSBoundParameters.ContainsKey('SecretName')) {
+        $Secrets = $SecretName.ToLower().Replace("_", "-")
+    }
+    else {
+        $Secrets = (Get-AzKeyVaultSecret -VaultName $_).Name
+    }
     $Secrets | ForEach-Object {
+        $OriginalSecretName = $_
         $SecretName = $_.ToUpper().Replace("-", "_").Replace("`"", "")
-        $SecretValue = (Get-AzKeyVaultSecret -VaultName $KeyVaultName -Name $_).SecretValue | ConvertFrom-SecureString -AsPlainText
-        $ContentType = (Get-AzKeyVaultSecret -VaultName $KeyVaultName -Name $_).ContentType  
+        if ($VersionHistory) {
+            $Versions = Get-AzKeyVaultSecret -VaultName $KeyVaultName -Name $_ -IncludeVersions
 
-        $SecretHash = @{
-            "SecretName" = "$SecretName"
-            "SecretValue" = "$SecretValue"
-            "ContentType" = "$ContentType"
+            $VersionHistoryHash = @()
+            $Versions | ForEach-Object {
+                $Version = $_.Version
+                $Updated = [TimeZoneInfo]::ConvertTimeBySystemTimeZoneId($_.Updated, 'Central Standard Time')
+                $SecretValue = Get-AzKeyVaultSecret -VaultName $KeyVaultName -Name $OriginalSecretName -Version $Version -AsPlainText
+                $VersionHash = @{
+                    "Version" = "$Version"
+                    "SecretValue" = "$SecretValue"
+                    "Updated" = [datetime]"$Updated"
+                }    
+                $VersionHistoryHash += $VersionHash
+            }
+            Write-Host "Retrieving last $VersionHistoryLength revisions of $SecretName..." -ForegroundColor DarkGray
+            Write-Host "[$KeyVaultName] Versions for ${SecretName}:" -ForegroundColor Green
+            $VersionHistoryHash | 
+                Sort-Object -Property Updated -Descending -Top $VersionHistoryLength |
+                Select-Object -Property Updated,SecretValue |
+                Format-Table -AutoSize
+            return
         }
-        $KeyVault += $SecretHash
+        else {
+            $SecretValue = Get-AzKeyVaultSecret -VaultName $KeyVaultName -Name $_ -AsPlainText
+            $ContentType = (Get-AzKeyVaultSecret -VaultName $KeyVaultName -Name $_).ContentType  
+
+            $SecretHash = @{
+                "SecretName" = "$SecretName"
+                "SecretValue" = "$SecretValue"
+                "ContentType" = "$ContentType"
+            }
+            $KeyVault += $SecretHash
+        }
     }
     $KeyVaults | Add-Member -NotePropertyName $KeyVaultName -NotePropertyValue $KeyVault
+}
+
+if ($VersionHistory) {
+    exit 0
 }
 
 Add-Content -Path "${File}.tmp" -Value ($KeyVaults | ConvertTo-Json)
